+++ lib/utils/global_enum.dart
@@ -1,0 +1,9 @@
+enum OrderType {
+  dineIn,
+  takeaway
+}
+
+enum PaymentType {
+  cash,
+  sgqr
+}
